fixed some junits tests

+++ src/main/java/edu/ntnu/idatt2001/erikskj/register/ArmyRegister.java
@@ -79,15 +79,18 @@ public class ArmyRegister {
         armies.remove(army);
 
         //gets file path
+        if (!army.isUploaded()){
+            File file = new File(army.getArmyFile().getPath());
+            //tries to delete file
+            try{
+                file.delete();
+            }
+            catch (Exception e){
+                System.err.println(e);
+            }
+        }
         /*
-        File file = new File(army.getArmyFile().getPath());
-        //tries to delete file
-        try{
-            file.delete();
-        }
-        catch (Exception e){
-            System.err.println(e);
-        }
+
 
          */
     }
